Move JSON type tests to their own package so they can be referenced by Dolt tests.
